Move useIsHydrated to x-internals package
