+++ packages/x-charts/src/ChartsYAxis/ChartsYAxisImpl.tsx
@@ -3,6 +3,7 @@ import * as React from 'react';
 import clsx from 'clsx';
 import useSlotProps from '@mui/utils/useSlotProps';
 import { useThemeProps, useTheme, styled } from '@mui/material/styles';
+import { useIsHydrated } from '@mui/x-internals/useIsHydrated';
 import type { ChartsYAxisProps, ComputedAxis, ScaleName } from '../models/axis';
 import { ChartsSingleYAxisTicks } from './ChartsSingleYAxisTicks';
 import { ChartsGroupedYAxisTicks } from './ChartsGroupedYAxisTicks';
@@ -10,7 +11,6 @@ import { ChartsText, type ChartsTextProps } from '../ChartsText';
 import { defaultProps, useUtilityClasses } from './utilities';
 import { isInfinity } from '../internals/isInfinity';
 import { useDrawingArea } from '../hooks/useDrawingArea';
-import { useIsHydrated } from '../hooks/useIsHydrated';
 import { isOrdinalScale } from '../internals/scaleGuards';
 import { getStringSize } from '../internals/domUtils';
 import { AxisRoot } from '../internals/components/AxisSharedComponents';
