+++ packages/x-chat-headless/src/message/MessageMeta.tsx
@@ -2,8 +2,8 @@
 import * as React from 'react';
 import useSlotProps from '@mui/utils/useSlotProps';
 import { SlotComponentProps } from '@mui/utils/types';
+import { useIsHydrated } from '@mui/x-internals/useIsHydrated';
 import { useChatLocaleText } from '../chat/internals/ChatLocaleContext';
-import { useIsHydrated } from '../chat/internals/useIsHydrated';
 import { ProgressIndicator, ProgressRoot, ProgressTrack } from '../internals/ProgressSlots';
 import { useMessageContext } from './internals/MessageContext';
 import { type MessageMetaOwnerState } from './message.types';
